Better inline pcg_rand() with constant limits
When pcg_rand() is passed a constant limit, two improvements are
possible:

  * the reject threshold can be calculated at compile time, so there's
    no need to make an effort to be nearly-divisionless

  * a resampling loop with a fixed threshold can be smaller than a
    fast path with an extern call to a slow path

So, change the macro that detected constant power-of-two limits so it
now detects any constant limit. When the limit is constant we use a
simpler inline rejection loop that calculates the reject threshold up
front. This gives the compiler enough information to eliminate the
loop for power-of-two limits, so that optimization still works.

The fast path / slow path is now only used for limits that vary at run
time.

The commentary explaining the rejection loop is now all in pcg.h so
there's less of an awkward split in the narrative. A few minor
improvements to the commentary elsewhere. Give more credit to Melissa
O'Neill.

Be more systematic about inline _fast and extern _small variants.
Automatically select which to use based on the compiler's optimization
flags.