Get rid of pcg_ulong_biased() and preen the commentary
It mainly existed because lines are not all long enough to fit
`(pcg_ulong_t)(pcg_random_fast(rng)) * (pcg_ulong_t)(limit)`.
But we can drop one of the casts and rely on implicit coercions
instead, which solves the problem. And when that expression is
inlined it makes more sense to collect the commentary together.